[ALPS07252696] mdla: fix coverity
fix overflow_before_widen

MTK-Commit-Id: a15b4177abc6a6195c226369641f65112630176d

CR-Id: ALPS07252696
Feature: apu_mdla_kernel

+++ drivers/misc/mediatek/apusys/mdla/new_arch/platform/v1_x/mdla_v1_x.c
@@ -100,7 +100,7 @@ static bool mdla_check_cmd_valid(
 		return false;
 	}
 	out_size = out_end - (uint64_t)ce->kva;
-	if (out_size < ce->count*MREG_CMD_SIZE) {
+	if (out_size < (uint64_t)ce->count*MREG_CMD_SIZE) {
 		mdla_err("%s:%d invalid cmd buffer size\n", __func__, __LINE__);
 		return false;
 	}
